Send notifications to conf organizers on signed digital contracts
When a digital contract is signed, notifications were only sent from
signwell to the actual party signing - not to the conference organizers.
So there was no way for them to know. With this change, also notify the
conference organiers on signing and counter-signing (counter-signing
only if automatic processing is n't enabled - if it is, the
notification is sent from there).