Implement basic feature for "favorite sessions"
This allows an attendee to mark which session are their favorites in a
way to "build a personal schedule". UI idea is the standard "stark-mark
some sessions". No actual UI is included at this point though, it's juat
an API endpoint that's to be used from the frontend websites. Given
nobody (hopefully) uses the old-style scheule that's in the core repo
there's not much point to build an implementation there, so we'll build
it elsewhere and then eventually backport it into the system once we
have a complete example.

This prodvides the foundation for collecting the information and for
viewing it in the backend (there's a report for top session and the
session count shows up in the schedule builder).

Starts work on #89, but it's far from complete as it's still pending
custom ical, generating notifications etc.