Track and show processing dates of digital contracts
Instead of just keeping track of that a contract has been signed,
explicitly track *when* it was sign. Also track when a contract is
partially signed (for confsponsor that's signed by the sponsor but not
by the organizers). And show this information on the sponsorship
dashboard and the sponsorship details.