Implement "additional contracts" for sponsors
This allows the management of additional contracts for each sponsor,
such as training contracts. If configured, this integrates with the
digital contract signing system, making it possible to handle these
contracts digitally as well. And if not, it still gets contract
templating the same way as the main sponsor contracts.