Fix copying of sponsorship levels with sponsor sessions
Since we store the track *id* in the database (not like regtypes where
we store the regtype *name*) it was impossible to copy a sponsorship
level that had benefits of this class. Both because we were looking them
up by name (but using the id as the name) and because a track with the
same *id* will never exist in a different conference.

Fix this by introducing a transform function that's called during the
copying to transform any parameters, and then implement that function
for sponsorsessions to look up a track with the same name and map to
that one.

Reported by Daniel Gustafsson