Add the option queryset_extra_columns to backend lists
This adds a field to the .only() part that tells the django ORM which
fields to fetch. It's needed when we a columns is indirectly used to
display a field, but not actually included. Example of this is the
errorcount on tweets, which is now also added by this commit.