Re-implement the braintree javascript frontend
Braintree discontinued the version we were using (apparently told people
about that many ages ago, but one has to go read their documentation to
find out, it's not pushed anywhere).

Their documentation is woefully incomplete and contradictive, but with
some random copy/paste from partially incorrect examples, this does seem
to work. And it's better than before.

I *think* we are now using the javascript api v3 version 3.97.3 (even
though their official releases list 3.96.0 as the latest) version 1.40.2
(yes, they have at least three different version numbers).

Unfortunately this version number is hardcoded in the javascript because
there is no way to get "latest compatible version" anymore (at least not
documented). So any users of braintree needs to check back to their
examples in the documentation regularly to see if they suddenly have put
out a new version.

I'm hesitant to call braintree ready for production based on how they do
things, and as a result of that the same applies to our support for
braintree. But it does seem to work for now.