Implement checkin field-scanning
The idea behind this is to add secondary "stations" at a conference
checkin which can off-load the main check-in. Typical example can be a
desk for handing out t-shirts separate from the registration.

Each checkin-field is tied to one of the dynamic reporting fields
implemented earlier, and once the "store" button is clicke on the
checkin dialog, the date of it being done is stored in the field.

Unlike regular check-in, this the check-in field scanning is done off
the *badge* not the ticket. This makes it more like the sponsor badge
scanning, and therefor we have to add a new landing page that can handle
the case when the same person is both a sponsor scanner and a check-in
scanner. It will then just prompt the user for which one.

This merges the two separate but extremely similar web scanner apps used
for checkin and badge scaninng into one, that is driven by parameters.
There's enough complexity there that we don't want two independent
ones...

In passing, support for old-style tokens (prior to 074e324b, almost a
year ago) is removed to simplify the code. That means that tickets or
badges issued before 074e324b was deployed can no longer be scanned, but
given how long ago it is this should not happen.

Per original idea from Vik Fearing with input from several more people
over time.