Publish a .well-known/assetlinks.json
This is required (in some cases? Because many seem to work) to publish
deep links on modern android versions. Basically we tie in the
publishing certificate we use in the play store to our domain, thereby
allowing it to "capture" those links directly.

Per complaints from the play console, and their suggested fix.