Switch badge scan output to use comma separator
We used semicolon separator to need less quoting, but it seems Excel
can't handle multiline strings in this case (google docs and libreoffice
both work, but badge scanning consumers are much more likely to use
Excel). This switches it to plain "excel dialect" in the python csv
module, which hopefully would be compatible with excel.