Fix uniqueness for conference session feedback
We need a unique constraint on (conference, session, attendee), so each
attendee can't end up with more than one entry. This can't be done in
the UI, but with a quick reload or a network blip could happen.

We also need to wrap the view in a transaction to avoid it happening in
the future (but there are still cases where it can happen, so we want both this
and the index)