Further attempts at validating migrations
This requires that we have secret_key set and all the dependencies
present. So use the dev_requirements.txt file to install all
dependencies. Except pillow which doesn't have a wheel for the
version we use, so just use the latest and greatest for that one
(since we won't actually call it, we just need the import to
 succeed, it doesn't matter that it's an incompatible version).