Remove dependency on bootstrap collapse for sponsor signup and admin
The bootstrap collapse works different under different versions of
bootstrap, and this caused issues for conferences using bootstrap 3 on
their site-wide skin (only site-wide skins are used for the sponsor
section, not per-conference ones, which is probably why this one went
missed so long). Instead, rely on <details> and <summary> which are in
HTML and have been in major browsers for >10 years.

Fix this on the backend side as well even if it's not really needed
there, so we stick to the same format on frontend and backend.

Thanks to Vik for a pointer in the right direction in that this is
possible.