Support storing video links on sessions
For recorded sessions, this lets us store things like youtube video ids,
later to be used on the schedule.

To store this, a jsonb object is added to each session. The list of
available providers can be configured on the conference, and a field to
enter the link data appears automatically. No validation of the data is
supported, so care has to be taken when adding. More logic around this
can be added later.

In passing, update the backend form for sessions to use field sets to
look nicer, and thereby re-ordering a few fields to match.