Trap conference template errors and avoid crashing
When reandering a conference template (from the conference specific
repo/directory), and a template exception is thrown, instead of crashing
with a stackdump and a 500 error to the user, trap the exception and
render the upstream template instead. The user won't get the conference
specific skin, but they will get a working page. To assist debugging, if
this happens also add a http header X-Conference-Template-Error to
indicate this is what happened.