Implement automatic inclusion of hashtags
With the increased importance of hashtags in the Mastodon world, make
the support more efficient and useful. This adds the ability to flag a
hashtag as "auto add", in which case it will be automatically added at
the end of all posts giong out to twitter/mastodon as a result of news
posts (this was not possible at all before). Additionally, it will also
get pre-filled in on manual posts made both using the backend interface
and the mobile webapp.

To control the order of appearance of the hashtags, a sortkey field has
been added to the table as well. Existing hashtags will just be sorted
alphabetically.