Disallow renaming regtypes referenced in sponsorsihp benefits
We already had protection in place to prevent it from being deleted,
but it was possible to rename a regtype when it was referened in a
sponsorship benefit which would then cause a crash when trying to claim
it. This canont be enforced with a foreign key as it's set inside a json
object, so we need an explicit check for it.