Handle database level constraint errors in backend editor
Backend editor would just crash if we failed to save due to a database
level constraint (foreign key, unique etc), even if said constraint was
defined in django (not for foreign keys as they drive dropdowns, but
django does not validate uniqueness, and there could always be a CHECK
constraint as well). So trap any such exceptions when trying to save and
just show the full error message to the user, instead of crashing.