Prevent infinite recusion for unskinned conference on skinned sites
If jinja templates were disabled on a conference but there existed
templates at the system skin level *and* those skins used recursive
template lookups, we could end up in infinite recursion. As long as they
didn't use that or if skinning was enabled for the individual
conference, the problem would not occur.