Remove jquery, selectize and bootstrap js from conference frontend
This was used for the call for papers form handling of multiple
speakers at this point, and caused issues with skins that didn't realize
this and thus didn't include them.

Replace it with a native js version, which instead of doing autocomplete
just adds a button for "Add speaker" to add secondary speakers. This
shows a browser popup asking for the email address of an existing
profile.

In practice I think this actually makes it more clear to the user what's
going on than the previous autocomplete version, just not as pretty.

It was also used for tags, which being a much shorter list, could just
be replaced with a set of checkboxes. Tags was also something most
conferences didn't use, and it only showed up if explicitly enabled on
the conference (non-default).