Move frontend javascript to separate file
Move all the different javascript snippets that were sitting in
individual pages into a shared file, that is then loaded from those
pages that need it (meaning it will end up being cached). And it's
generally considered "better" not to have inline scripts.

This specifically does *not* include changes to the "view feedback"
parts which are pulling in google scripts and doing all sorts of more
complicated inline js. That deserves a separate proper fix.

Requires all skins to properly define the `pagescript` block, but tihs
was already a requirement, it just changed how it looks if it's not
there.