Replace Google charts javascript with local SVG charts
These charts are (of course) not quite as advanced as the Google ones,
but avoids an external depdendency on Google and indeed the local js
completely. Also makes for a much faster rendering of the pages with
feedback.

This leaves the admin-only time graph as the only jsapi depdendency on
google, to be handled later.