Replace google js charts with svg for time graphs as well
This removes the final dependency on google js apis.

In passing also make it possible to filter the conferences when picking
some, as that has become increasingly annoying as the number of
available conferences grow.

This also removes the non-linear trendlines support as that has never
really been useful anyway. Linear trendlines remain.