Separate the message to sponsor and org in digital contracts
With an extension of the upstream APIs we can now send separate mssages
to the different parties of the signing. With this we can keep the
"sponsor friendly" message we have now for the sponsors, and send a more
appropriate message to the organizers to make it clear what the next
step is.

In the internal API we make it an optional parameter, so if another
future provider doesn't support separating the messages, things still
work.