Explicitly stop hyphenating email address
In some unfortunate circumstances, the email address of the conference
was hyphenated on the registration dashboard making it seem like it is
spelled differently from the actual address. This page is riddled with
inline styles so add an inline for this as well, saving a more thorough
CSS cleanup for another patch.