Allow fetching twitterpost thumbnails publicly
In doing so also add the existance of an image as a hasimage key in the
list of posts. The idea is this can be used to show the thumbnails as
part of a post listing. We don't allow local access to the full images -
if we want to do that in the future, we probably need to consider some
form of outside cache as they're pretty big objects in the database.