Poliy and attachment urls should work with and without trailing slash
We always send out the link including the trailing slash, but apparently
either some MUAs or browsers remove the trailing slash. In this case our
rule would not match. And since we have legacy urls that include the
register/ part, they get captured by that rule (with the conference name
being "policy" or "attach") instead of the django default rule.

So, update the regexp to allow these links with or without the trailing
slash.