Add function to send email to arbitrary external recipients
Allow sending of simple short emails to external addresses
with the sender set to one of the conference addresses (contact,
sponsors, contract). This makes it possible to send email from the
"official addresses" which will be sent from the appropriate servers
with correct DKIM etc, making it pass spam filters more easily.