Fix off-by-one-regexp in URL matching for speakers
We're not suppose to allow / in the speaker name, as that can cause
infinite recursions. But due to an extra . in the regexp, if we loaded
the speaker from the schedule (not the session list) we could, which if
you had a bad link somwehere a bot could then end up traversing into
without getting a 404.