Implement multi-claim sponsorship benefits
This basically allows a single benefit to be claimed multiple times with
different data, for benefit classes that support it (which is most but
not all). This can be used for things like "two sessions in the sponsor
track" or "send two different pieces of merch", to avoid the previously
used workarounds of creating mutiple benefits with slightly different
names.