Add support for token-downloads of sponsor level metadata
In passing, this adds support for YaML format downloads of all token
based data.

The token data will generate json or yaml with information about all
sponsorship levels and benefits created in the system. The idea being
this can be pulled into a context file and thereby guaranteed to be the
same in the static and dynamic parts of the site. We'know from
experience these can easily get slightly out of sync.