Add support for token download of sponsor claimed benefits
This allows downloading a json/yaml file with all the claimed benefits
(including information on if they are confirmed or not).

This also adds a tool called fetch_sponsorfiles.py that will parse this
data and be able to download related files, such as "Logo for the
website" benefits.

The details being provided probably need some more going over with some
reasl world experience, so for now this should be considered
experimental.