Activate the conference timezone when fetching sponsor object
When using _get_sponsor_and_admin() to get the information we'd leave
the timezone at the system default, which is not correct when a
conference is run in a different timezone from the main one on the
server. Timezones haven't really mattered to sponsor info before (more
or less only the list of emails would be affected), but now that we have
deadlines configured it must be done.