Add sessionid attribute to cross conference sessions
These sessions also have an id, and we want to expose it to the
schedule. We cannot do that using 'id' now as there are many schedule
templates out there that use the lack of 'id' to determine if a session
is cross conference or not, so we create a new one.