Send notification when secondary speaker is added during cfp
When speaker A adds speaker B as a co-speaker of a session, send a
notification to speakear B about this so they're not surprised by it.

This only happens during the cfp process, not if admins make changes in
the backend (and per that it cannot happen after the cfp has closed).

Suggested by Karen Jex