Don't try to post empty content on multiprovider tweets
When rendering multiprovider tweets, if there is no messaging at all
defined, we get no text, and that's normal. However, when then also need
to check for no text at all in the posting side of things *before* we
check for providers, or we will violate a database constraint.

Of course, one should define tweet templates when there is no messaging
configuration, but we should also not crash if it's done.

Spotted by Karen Jex