Show social media post status per network even for single text posts
Previously we'd just show a boolean "posted" or "not posted" based on if
there was more than one version of the text. But it's actually
interesting to know if it's posted to only some of the  networks even if
the exact same text is posted to all of them, so make it show that in
all cases.