Replace a number of <input type="submit"> with <button type="submit">
This separates the value posted from the text on the button, which
*hopefully* will prevent the errors caused by people using browser
extensions that translate the site, thereby submitting buttons in german
and french for example.

This only touches buttons on the frontend side, at least for now.
Backend users are expected to not use such extensions. And we keep
setting the value to the same as before, so that if some skins have
completely overriden the forms and added their own HTML, it will
continue to work.

This will likely require some frontend skins to update their CSS rules
to match. Things will still work without such changes, but will not look
as good.

Hopefully fixes #150