Add support for qrcode library, and make it default
The qrencode library we've been using is basically dead (and has been
for a while), so add support for the newer pure-python qrcode library.

We keep support for the qrencode one around, so we don't break things
on existing installs, but consider it deprecated and should eventually
be removed.

Fixes #123