Expose that sponsor cost is a VAT field in backend interface
While normally sponsors are invoived without VAT, there are cases where
VAT is added and in those cases we want the same VAT preview there as we
have for registration types etc.

The code for dealing with that preview was clearly already designed to
handle this, but it was never enabled for levelcost.