Add special registration type requiring a voucher
ABy requiring a voucher to use a specific registration type, it can
become an "invite only" regtype. The cost set for it doesn't matter, it
can still be zero, but the attendee needs to be in posession of a
voucher to complete the registration. Not as convenient as using the
auto-approving ones like speaker or staff, but a lot more flexible.