Properly remove "confirm preview" checkbox on broken uploads
When an image uploaded as a sponsor benefit was broken (unparseable), we
would incorrectly still show the "confirm preview looks ok" checkbox
below the error message. If the user confirmed that the preview (which
was non-existant and replaced with an error message) looked OK, we would
store an empty image in the database and consider the benefit OK.

Instead, we're of course not supposed to show the preview checkbox at
all if the image uploaded is broken.