Attempt best-effort on transferwise transactions with no permissions
If a transferwise transaction is sent from another user of their bank,
the id of the transaction belongs to the *other user*, and the recipient
has no permissions to get to the details. So when this happens, do a
best-effort set of details based on what we can find in the activity
(which generally does not contain payment references for example), and
leave it to the user - rather than just ignoring it.

The same seems to happen for direct debit transactions, so process them
the same.