Add "check-in message" to registration types
This message will be shown (highlighed) when an attendee is checked in,
and can be used to for example alert the person doing the check-in that
they should validate the registration by checking a student-id.

To do so, the "highlight this field" functionality is moved into the
backend and returned as part of the API, instead of the very limited "if
' NOT ' is present in the field text then highlight it" (used only for
unconfirmed policies) that was previously used.

Fixes #190