Reimplement talkvote page in a more flexible way
This reimplements the talkvote page mostly client-side, making it
possible to do all the filtering without an expensive server roundtrip
(it gets expensive to do this due to the page being big for large
 conferences). Also implement column on/off checkboxes.

This also means getting rid of the bootstrap accordion controls, in
favor of native CSS, which as a bonus is faster.

Based on ideas from Karen Jex, and several years of discussions.

Fixes #193, along with many other things