Make transfer registration use ajax selectize for account names
Loading the transfer registration form actually loaded a list of every
single user account on the system into the HTML, which was very slow.
This is because it doesn't use the generic form interface, which means
the JS to set it up to load async was not present - so we now instead
make an explicit initialization.