Change stripe payments to do server side redirect
It makes little sense to load a separate page, which loads a third party
javascript, and then executes it inline just to perform a redirect to an
URL that's already in the response from the API.

It does lead to one extra (cheap) stripe api call if the user initates
the payment at stripe, doesn't complete it, and rstarts the process from
the invoice. But in all other (thus, the normal) cases it leads to less
calls, and avoiding to load 3rd party javascript inline in general is a
good thing.