Don't serialize full objects in ConcurrentProtectedModelForm
The idea behind the form is to serialize the state of the data from the
form before it's rendered and then make sure it hasn't changed before
it's submitted, since that would overwrite concurrent changes. The
problem with the implementation is it serialized the entire objects on
the other side of a foreign key or manytomany relationship, which could
be very large if it was for example Speaker records with photos. (One
example found made a simple form over 2.5Mb due to this, which triggered
the oversize form limits)

Update this to instead just serialize the primary key of the object. The
way forms work that *should* always exist at this point otherwise it
cannot be added. And this makes the form a *lot* smaller and faster.

Per bug report from Karen Jex