Add support for generic file upload sponsorship benefit class
This allows the upload of a generic file, not just an image. It can be
restricted by mime types if wanted (to for example say only EPS and SVG
files), and will then validate the MIME header. It will *not* attempt
any further validation beyond the MIME header. If no MIME type is
specified on the benefit, then it will accept any uploaded file formats
and leave it entirely to the administrator to handle.

To make this work, also add support to the InlineEncodedStorage to store
metadata, and use that to store the filename so we can reuse it later.