Actually allow incomplete registrations to b added to addopts
The fields allowed incomplete registrations to be added to the list of
users for an additional options, but the lookups that generated the
autocomplete lists did not include them, so in the end it couldn't
actually be used. To fix it we have to create a new lookup that includes
this set of users.

Spotted by Andreas Scherbaum