Add tool to generate images from schedule data
This is a trivial tool to generate images for each session in
the system. This can be used to for example generate screens
to show on monitors outside the conference rooms, or tor
generate slides to show in the rooms in between talks.

Based off an idea from Chris Ellis but reimplemented using jinja2
templates (and python instead of node.js) to use the same templates as
the rest of the system uses.